Updating weather integration on front page

+++ WeatherApi/WeatherData.cs
@@ -7,7 +7,7 @@ namespace WeatherApi
     public class WeatherData
     {
         public DateTime dt_txt { get; set; }
-        public decimal temp { get; set; }
+        public int temp { get; set; }
         public string icon { get; set; }
     }
 }
